+++ examples/example_current_from_drifter.py
@@ -35,7 +35,7 @@ o.add_reader(r)
 lons = [2.39, 2.391, 2.392, 2.393, 2.394, 2.393, 2.392, 2.391, 2.39]
 lats = [60.02, 60.02, 60.019, 60.02, 60.021, 60.022, 60.021, 60.021, 60.02]
 o.seed_within_polygon(lons=lons, lats=lats,
-                      number=1000, time=r.start_time)
+                      number=2000, time=r.start_time)
 
 #%%
 # Finally running simulation