+++ opendrift/scripts/hodograph.py
@@ -103,7 +103,7 @@ if __name__ == '__main__':
     plt.ylabel('North component [m/s]')
     plt.xlabel('East component [m/s]')
     textstr = u'%s UTC\n%.3fN\N{DEGREE SIGN}, %.3fE\N{DEGREE SIGN}' % \
-             (time, np.float(args.lat), np.float(args.lon))
+             (time, float(args.lat), float(args.lon))
     ax.text(0.05, 0.95, textstr, transform=ax.transAxes, fontsize=14,
             verticalalignment='top',
             bbox=dict(boxstyle='round', facecolor='wheat', alpha=0.5))